\name{getMetricFunctionMetadata}

\alias{getMetricFunctionMetadata}


\title{Return JSON Metadata for Metric Functions}

\description{
Returns a JSON formatted string with metric function metadata.
This string is needed by the python-based ISPAQ command-line utility developed
by IRIS DMC.
}

\usage{
getMetricFunctionMetadata()
}

%% \arguments{ }

%% \details{ }

\value{
JSON formatted string containing metric function metadata.
}

%% \references{ }

\author{
Jonathan Callahan \email{jonathan@mazamascience.com}
}

%% \note{ }

%% ~Make other sections like Warning with \section{Warning}{....} ~

%% \seealso{ }

