\name{metricList2DF}

\alias{metricList2DF}


\title{Convert a MetricList into a Tidy Dataframe}

\description{
The \code{metricList2DF} function converts a list of \code{SingleValueMetric}s into a
"tidy" dataframe with one value per row..
}

\usage{
metricList2DF(metricList)
}

\arguments{
  \item{metricList}{a list of \code{SingleValueMetric} objects}
}

\details{
Metrics functions return lists of \code{SingleValueMetric} objects.  A long \code{metricList} may be built up
by appending the results of different metrics functions or the same metrics function operating on different seismic signals.
A metricList generated by any of the MUSTANG Rscripts can be stored as an \code{.RData} file and reloaded for examination.

A metricList may contain values for many different metrics.  This function creates a single "tidy" dataframe with
the following colulmns: \code{metricName, value, snclq, starttime, endtime, qualityFlag}.
}

\value{
A dataframe with one row per metric measurement.
}

%% \references{ }

\author{
Jonathan Callahan \email{jonathan@mazamascience.com}
}

%% \note{ }

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{SingleValueMetric-class}}, 
\code{\link{metricList2Xml}}
}

\examples{
  \dontrun{
# Open a connection to IRIS DMC webservices
client <- new("IrisClient")

# Get the waveforms
starttime <- as.POSIXct("2012-01-24", tz="GMT")
endtime <- as.POSIXct("2012-01-25", tz="GMT")
st1 <- getDataselect(client,"AK","PIN","","BHE",starttime,endtime)
st2 <- getDataselect(client,"AK","PIN","","BHN",starttime,endtime)
st3 <- getDataselect(client,"AK","PIN","","BHZ",starttime,endtime)

# Calculate metrics and append them to the metricList
metricList <- stateOfHealthMetric(st1)
metricList <- append(metricList, basicStatsMetric(st1))
metricList <- append(metricList, basicStatsMetric(st2))
metricList <- append(metricList, basicStatsMetric(st3))

# Create dataframe
metricDF <- metricList2DF(metricList)
head(metricDF)
  }
}

